Expose tests to pytest
